feat: add metadata option
This commit introduces a new metadata option for the tool.
With this option enabled, a metadata.json file will be created inside
the target folder. The metadata.json file contains metadata obtained
using the "TelegramClient.getMessages" method.
The metadata.json file can be used for various purposes such as
analysis, filtering, or integration with other scripts.

Please note that this implementation is currently unoptimized and may
not perform efficiently for large arrays. Optimizations will be made
before an official release.